# api/NeoWallet_OpenAPI_v1.yaml
openapi: 3.1.0
info:
  title: NeoWallet API
  version: 1.0.0
  description: NeoWallet MVP API Contract Specification
  contact:
    name: NeoWallet API Team
    email: api@neowallet.com
  license:
    name: Proprietary
    identifier: Proprietary
servers:
  - url: https://api.neowallet.com/api/v1
    description: Production
  - url: https://api-staging.neowallet.com/api/v1
    description: Staging
  - url: https://api-dev.neowallet.com/api/v1
    description: Development
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      required:
        - code
        - message
        - requestId
        - timestamp
      properties:
        code:
          type: string
          example: BUDGET_NOT_FOUND
        message:
          type: string
          example: Budget was not found.
        requestId:
          type: string
          format: uuid
        timestamp:
          type: string
          format: date-time
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
    Pagination:
      type: object
      properties:
        page:
          type: integer
          minimum: 1
          default: 1
        limit:
          type: integer
          minimum: 1
          maximum: 100
          default: 20
        totalCount:
          type: integer
        pageCount:
          type: integer

    # Authentication Schemas
...
    Bill:
      type: object
      properties:
        billId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        familyId:
          type: string
          format: uuid
          nullable: true
        name:
          type: string
          maxLength: 100
        amount:
          type: number
          format: decimal
          exclusiveMinimum: 0
        currency:
          type: string
        dueDate:
          type: string
          format: date
        category:
          type: string
        isRecurring:
          type: boolean
        recurringPeriod:
          type: string
          enum: [MONTHLY, WEEKLY, YEARLY]
          nullable: true
        vendor:
          type: string
        notes:
          type: string
        status:
          type: string
          enum: [PENDING, PAID, OVERDUE]
        paidDate:
          type: string
          format: date
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateBillRequest:
      type: object
      required:
        - name
        - amount
        - dueDate
      properties:
        name:
          type: string
          maxLength: 100
        amount:
          type: number
          format: decimal
          exclusiveMinimum: 0
        currency:
          type: string
        dueDate:
          type: string
          format: date
        category:
          type: string
        isRecurring:
          type: boolean
          default: false
        recurringPeriod:
          type: string
          enum: [MONTHLY, WEEKLY, YEARLY]
        vendor:
          type: string
        notes:
          type: string
    UpdateBillRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 100
        amount:
          type: number
          format: decimal
          exclusiveMinimum: 0
        dueDate:
          type: string
          format: date
        category:
          type: string
        isRecurring:
          type: boolean
        recurringPeriod:
          type: string
          enum: [MONTHLY, WEEKLY, YEARLY]
        vendor:
          type: string
        notes:
          type: string
    MarkPaidRequest:
      type: object
      properties:
        paidDate:
          type: string
          format: date
        paymentMethod:
          type: string
        notes:
          type: string
    MarkPaidResponse:
      type: object
      properties:
        billId:
          type: string
          format: uuid
        status:
          type: string
          enum: [PAID]
        paidDate:
          type: string
          format: date
        paymentMethod:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        updatedAt:
          type: string
          format: date-time

    # Financial Health Schemas
...
paths:
...
  /bills:
    post:
      summary: Create bill
      operationId: createBill
      tags:
        - Bills
      security:
        - bearerAuth: []
      parameters:
        - name: X-Family-ID
          in: header
          required: false
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          required: false
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBillRequest'
      responses:
        '201':
          description: Bill created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Bill'
        '401':
          description: Invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Not family member
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    get:
      summary: Get bills with filtering
      operationId: getBills
      tags:
        - Bills
      security:
        - bearerAuth: []
      parameters:
        - name: X-Family-ID
          in: header
          required: false
          schema:
            type: string
            format: uuid
        - name: page
          in: query
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - name: filter[status]
          in: query
          schema:
            type: string
            enum: [PENDING, PAID, OVERDUE]
        - name: filter[category]
          in: query
          schema:
            type: string
        - name: filter[dueDateFrom]
          in: query
          schema:
            type: string
            format: date
        - name: filter[dueDateTo]
          in: query
          schema:
            type: string
            format: date
        - name: sort
          in: query
          schema:
            type: string
            example: dueDate:asc
      responses:
        '200':
          description: Bills retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  bills:
                    type: array
                    items:
                      $ref: '#/components/schemas/Bill'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '401':
          description: Invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Not family member
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /bills/{billId}:
    get:
      summary: Get bill by ID
      operationId: getBill
      tags:
        - Bills
      security:
        - bearerAuth: []
      parameters:
        - name: billId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Bill retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Bill'
        '401':
          description: Invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Not authorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Bill not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      summary: Update bill
      operationId: updateBill
      tags:
        - Bills
      security:
        - bearerAuth: []
      parameters:
        - name: billId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateBillRequest'
      responses:
        '200':
          description: Bill updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Bill'
        '401':
          description: Invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Not authorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Bill not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      summary: Delete bill
      operationId: deleteBill
      tags:
        - Bills
      security:
        - bearerAuth: []
      parameters:
        - name: billId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Bill deleted successfully
        '401':
          description: Invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Not authorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Bill not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /bills/{billId}/mark-paid:
    post:
      summary: Mark bill as paid
      operationId: markBillPaid
      tags:
        - Bills
      security:
        - bearerAuth: []
      parameters:
        - name: billId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          required: false
          schema:
            type: string
            format: uuid
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MarkPaidRequest'
      responses:
        '200':
          description: Bill marked as paid successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MarkPaidResponse'
        '401':
          description: Invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Not authorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Bill not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Invalid status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # Financial Health Paths
...